+++ build.sh
@@ -11,5 +11,6 @@ fi
 dotnet test SentryNoSamples.slnf -c Release -l $testLogger \
     /p:CollectCoverage=true \
     /p:CoverletOutputFormat=opencover \
+    /p:CopyLocalLockFileAssemblies=true \
     /p:Exclude=\"[Sentry.Protocol.Test*]*,[xunit.*]*,[Sentry.Test*]*\" \
     /p:UseSourceLink=true